feat(favoris): deux modes de livraison Essentiel ⇄ Flâner, sheet unifiée
Story 10.2 — mobile/UX uniquement, aucune modif backend/Alembic.

Avant, une source mise en favori apparaissait DANS LES DEUX chemins de
rendu (section Essentiel *et* onglet Flâner) faute de notion de « mode ».

- Appartenance exclusive sans migration : une source est livrée dans
  l'Essentiel (Tournée, cap 5) ssi sa clé `source:<id>` est dans
  `tournee_order_v1` ; sinon elle vit en Flâner (onglets, cap 10).
- Règle d'appartenance centralisée sur `TourneeOrderState`
  (`sourceIsEssentiel` / `essentielSourceKeys`) — une seule source de
  vérité partagée par le provider Tournée, les onglets Flâner et la sheet
  (évite la dérive entre les chemins).
- Provider : `_pickFavoriteSources` filtre par mode + le listener refetch
  quand l'ensemble des clés `source:` change (entrée/sortie de l'Essentiel).
- Onglets Flâner : exclusion des sources mode-Essentiel.
- Sheet unifiée `manage_favorites_sheet.dart` (deux portes : Essentiel /
  Flâner) ; déplacement de mode explicite par source ; funnel veille 🔭
  sur les sujets. `showTourneeComposerSheet` / `showPinSubjectsSheet`
  deviennent des shims vers `showManageFavoritesSheet`.

Vérif : flutter test ciblés (67) + tournee_order_prefs verts ;
flutter analyze clean sur les fichiers touchés. Suite mobile complète =
24 échecs pré-existants hors-scope (Hive/Supabase non init), identiques
sur l'arbre de base ; CI = pytest backend only.
