+++ docs/stories/core/10.2.favoris-modes-livraison.md
@@ -1,0 +1,102 @@
+# Story 10.2 — Favoris : deux modes de livraison (Essentiel ⇄ Flâner)
+
+> Type : **Feature** (mobile/UX uniquement, aucune modif backend/Alembic) · 1 PR vers `main`.
+> Suite de [10.1 — Sujets épinglés Flâner](10.1.sujets-epingles-flaner.md).
+
+## Problème
+
+Deux mécaniques de favoris se chevauchaient sans frontière claire : la curation
+**Tournée du jour** (favoris → sections de l'Essentiel) et l'**épinglage Flâner**
+(→ onglets de navigation). Symptôme vérifié : **une source mise en favori
+apparaissait DANS LES DEUX** (section Essentiel *et* onglet Flâner), car aucun des
+deux chemins de rendu ne filtrait par un « mode ».
+
+## Décision produit
+
+1. Un seul axe de favoris, **deux modes de livraison** : « Chaque jour dans
+   l'Essentiel » (Tournée, cap 5) vs « En continu dans Flâner » (onglets, cap 10).
+2. Mode **exclusif** : chaque source favorite vit dans **une seule** section.
+3. **Sujets précis** restent Flâner-only. La voie « un sujet chaque jour » =
+   le convertir en **Veille** (funnel 🔭), pas construit ici, juste amorcé.
+4. MVP mobile, 1 PR.
+
+## Mécanisme — `tournee_order_v1` fait autorité pour les sources
+
+Plutôt qu'une 3ᵉ map ou une migration one-shot fragile (timing au boot), on
+choisit une règle de partition **sans migration** :
+
+- **Essentiel** revendique une source **ssi** sa clé `source:<id>` est dans
+  `tournee_order_v1` (`tourneeOrderPrefsProvider.order`).
+- **Flâner** revendique une source **sinon** (défaut). Donc :
+  - source en aucune pref → Flâner (anti-disparition automatique) ;
+  - source dans les deux (cas legacy) → Essentiel gagne (la présence dans
+    `tournee_order_v1` l'emporte), la clé résiduelle dans `pinned_tabs_order_v1`
+    n'affecte que l'ordre et est inoffensive.
+
+`pinned_tabs_order_v1` reste **ordre-only** pour les onglets (sujets + sources).
+Aucune migration n'est nécessaire : l'appartenance se déduit d'un seul signal.
+
+### Changements de rendu
+
+1. `flux_continu_provider._pickFavoriteSources()` — filtre les sources favorites
+   à celles dont `tourneeSourceKey(id) ∈ tournee_order_v1` avant le cap.
+2. Listener `tourneeOrderPrefsProvider` — quand l'ensemble des clés `source:` de
+   `order` change (une source entre/sort de l'Essentiel), déclenche
+   `_refetchSourcesOnly` (sinon la section n'existe pas encore dans `_sources`).
+3. `favorite_topic_tabs._buildTabModels()` — exclut une source dont la clé est
+   dans `tournee_order_v1` (mode Essentiel). Sujets inchangés.
+
+## Sheet unifiée — une sheet, deux portes
+
+`manage_favorites_sheet.dart` : `showManageFavoritesSheet(ctx, {entry})` (enum
+`essentiel`/`flaner` → pilote le segment « Ajouter » présélectionné ; contenu
+identique des deux côtés). Deux sections labellisées :
+
+- **« Chaque matin dans ton Essentiel »** (cap 5) — éditorial + thèmes + veille
+  + sources mode-Essentiel. Reorderable.
+- **« Tes onglets pour explorer »** (cap 10) — sujets + sources mode-Flâner.
+  Reorderable.
+
+Mouvement de mode : chaque ligne **source** porte une affordance « → Flâner » /
+« → Essentiel » (déplacement explicite, robuste — pas de drag cross-list entre
+deux `ReorderableListView`, non supporté nativement par Flutter). Les
+thèmes/veille (ancrés Essentiel) et sujets (ancrés Flâner) ne se déplacent pas
+entre sections ; chaque ligne **sujet** porte un 🔭 « chaque jour ? » → flow
+Veille (`RouteNames.veilleConfig`).
+
+Caps : Essentiel 5 (`kTourneeVisibleCap`), Flâner 10 (`kMaxFavoriteTabs`), trait
+« Hors … » dans chaque section. Cap backend sources = 5 **partagé** entre les
+deux sections (une source Flâner consomme le quota).
+
+Les deux points d'entrée existants deviennent des **shims** :
+`showTourneeComposerSheet` → `entry: essentiel`, `showPinSubjectsSheet` →
+`entry: flaner`. `PinSubjectsBanner` et le `+`/engrenage de `feed_filter_bar`
+re-pointés.
+
+## Fichiers touchés
+
+- `apps/mobile/lib/features/flux_continu/providers/flux_continu_provider.dart`
+- `apps/mobile/lib/features/feed/widgets/favorite_topic_tabs.dart`
+- `apps/mobile/lib/features/flux_continu/widgets/manage_favorites_sheet.dart` (nouveau)
+- `apps/mobile/lib/features/flux_continu/widgets/tournee_composer_sheet.dart` (shim)
+- `apps/mobile/lib/features/feed/widgets/pin_subjects_sheet.dart` (shim + banner)
+- `apps/mobile/lib/features/feed/widgets/feed_filter_bar.dart` (re-pointage)
+
+## Tâches
+
+- [x] Filtre exclusif sources (provider Tournée + listener)
+- [x] Filtre exclusif sources (onglets Flâner)
+- [x] Sheet unifiée + shims + re-pointage
+- [x] 🔭 funnel veille sur lignes sujet
+- [x] Caps 5/10 par section
+- [x] Tests + flutter analyze
+
+## Vérification
+
+- `cd apps/mobile && flutter test <ciblés> && flutter analyze`.
+- Pas de migration Alembic (1 head inchangé).
+
+## Note (suivi, hors PR)
+
+Sujets dans l'Essentiel via moteur Veille généralisé 1→N : PR backend séparée
+réutilisant `feed_filter.py`. Le funnel 🔭 en est l'amorce produit.
